Updated front-page.php to work with content-front.php
In class I had front-page.php working with content-home.php but this is
confusing and in the future we will make home.php so this change brings
things in order and makes more sense. content-front just shows a small
amount of info, there will be a featured image added eventually.
content-home.php will be used to show excerpts of posts on home.php
which will be the template for the blog page.