added 4.8.14 files  and foundation-wordpress
The 4.8.14 files are the ones we worked on in class that day and are
doing rapid HTML/CSS only prototypes. This shows some more things than
we did in class and gets the Orbit slider working. To get that working
more JavaScript files also had to be added.

the foundation-wordpress will be the base files we use to teach how to
take the prototypes in to a WordPress theme.